deleted feature auth. fixed problem use features in others
