+++ app/chat/[id]/page.tsx
@@ -11,13 +11,12 @@ import { Toaster } from "@/shared/ui/toaster"
 import "katex/dist/katex.min.css"
 import  {Markdown} from "@/features/markdown-renderer/ui"
 import MessageItem from "@/entities/chat/ui/MessageItem"
-import { useAuth } from "@/features/auth/model"
 import { useChats } from "@/entities/chat/model"
 import { scrollToBottom } from "@/shared/utils"
 import { useLockBodyScroll, useScrollVisibility, useAutoScroll } from "@/shared/hooks"
 import { useChatInitialization } from "@/features/chat-init/model"
 import { ThemeToggle } from "@/shared/ui/theme-toggle"
-import { useMessageContext } from "@/shared/contexts"
+import { useAuth, useMessageContext } from "@/shared/contexts"
 import { useUser } from "@/shared/contexts"
 import { useRouter } from "next/navigation"
 import { getToken } from "@/shared/utils"
