+++ app/profile/page.tsx
@@ -12,8 +12,7 @@ import ProfileSettings from "@/widgets/profile/profile-settings"
 import Subscription from "@/widgets/profile/subscription"
 import ProviderChoice from "@/features/provider-choice/ui/provider-choice"
 import Tarifs from "@/features/pay-for-plan/ui/tarifs"
-import { useAuth } from "@/features/auth/model"
-import { useUser } from "@/shared/contexts"
+import { useAuth, useUser } from "@/shared/contexts"
 import { useRouter } from "next/navigation"
 import { getToken } from "@/shared/utils"
