+++ entities/chat/model/use-chats.ts
@@ -1,10 +1,9 @@
 
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { getChatAllApi, getChatByIdApi } from './api';
-import { useUser } from '@/shared/contexts';
+import { useAuth, useUser } from '@/shared/contexts';
 import { ChatHistory } from './types';
 import { useParams, useRouter } from "next/navigation"
-import { useAuth } from '@/features/auth/model';
 import { clearChatApi } from '@/features/clear-chat/model/api';
 import { useMessageContext } from '@/shared/contexts';
 import { useMessage } from '@/entities/message/model';
