+++ features/auth/model/api.ts
@@ -1,5 +1,5 @@
-import { apiClient } from '@/shared/api';
+// import { apiClient } from '@/shared/api';
 
-export const loginApi = async (email: string, password: string) => {
-    return apiClient.post(`user/login`, { email, password });
-};
+// export const loginApi = async (email: string, password: string) => {
+//     return apiClient.post(`user/login`, { email, password });
+// };
