+++ features/auth/model/index.ts
@@ -1,4 +1,2 @@
-export * from './api'
 export * from './types'
-export * from './use-auth'
 export * from './use-login-form'
