+++ features/chat-init/model/index.ts
@@ -1,1 +1,2 @@
 export * from './use-chat-init'
+export * from './types'
