+++ features/chat-init/model/types.ts
@@ -1,0 +1,6 @@
+
+export interface UseChatInitializationProps {
+  isLoading: boolean;
+  setIsLoading: React.Dispatch<React.SetStateAction<boolean>>;
+  ws: React.MutableRefObject<WebSocket | null>;
+}
