+++ features/updatePassword/model/use-updatePassword.tsx
@@ -1,7 +1,6 @@
 
-import { useAuth } from '@/features/auth/model';
 import { updatePasswordApi } from './api';
-import { useUser } from '@/shared/contexts';
+import { useAuth, useUser } from '@/shared/contexts';
 export const useUpdatePassword = () => {
 
     const { setIsAuthenticated } = useAuth()
