+++ shared/contexts/index.ts
@@ -1,2 +1,3 @@
 export * from './MessageContext'
 export * from './user-context'
+export * from './use-auth'
