+++ shared/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from './use-auth'
+export * from '../contexts/use-auth'
 export * from './use-lock-body-scroll'
 export * from './use-media-query'
 export * from './use-mobile'
