+++ widgets/navigation/nav-links.tsx
@@ -1,9 +1,9 @@
 "use client"
 
+import { useAuth } from "@/shared/contexts"
 import Link from "next/link"
 import { usePathname } from "next/navigation"
 import { useEffect, useState } from "react"
-import { useAuth } from "@/features/auth/model"
 
 export const NavLinks = () => {
   const { isAuthenticated } = useAuth()
