+++ widgets/user-menu/user-menu.tsx
@@ -4,9 +4,8 @@ import { Avatar, AvatarFallback } from "@/shared/ui/avatar"
 import { Button } from "@/shared/ui/button"
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator, 
   DropdownMenuTrigger} from "@/shared/ui/dropdown-menu"
-import { useAuth } from "@/features/auth/model"
 import { Logout } from "@/features/logout/ui"
-import { useUser } from "@/shared/contexts"
+import { useAuth, useUser } from "@/shared/contexts"
 import { LogOut, User } from "lucide-react"
 import Link from "next/link"
 import { useRouter } from "next/navigation"
